simple-rag-backend/requirements.txt
fastapi
uvicorn[standard]
cohere
qdrant-client
python-dotenv
requests
beautifulsoup4